speech_to_text/requirements.txt
streamlit>=1.24.0
openai-whisper>=20231117
pyaudio>=0.2.11
numpy>=1.21.0
torch>=1.9.0
torchaudio>=0.9.0